refactor: Fake 결제 게이트웨이를 위한 추상화
